+++ hutool-aop/src/main/java/cn/hutool/aop/interceptor/CglibInterceptor.java
@@ -8,10 +8,16 @@ import cn.hutool.core.exceptions.UtilException;
 import net.sf.cglib.proxy.MethodInterceptor;
 import net.sf.cglib.proxy.MethodProxy;
 
-public class CglibInterceptor implements MethodInterceptor{
+/**
+ * Cglib实现的动态代理切面
+ * 
+ * @author looly
+ *
+ */
+public class CglibInterceptor implements MethodInterceptor {
 	private Object target;
 	private Aspect aspect;
-	
+
 	/**
 	 * 构造
 	 * 
@@ -22,28 +28,28 @@ public class CglibInterceptor implements MethodInterceptor{
 		this.target = target;
 		this.aspect = aspect;
 	}
-	
-	public Object getTarget(){
+
+	public Object getTarget() {
 		return this.target;
 	}
 
 	@Override
 	public Object intercept(Object obj, Method method, Object[] args, MethodProxy proxy) throws Throwable {
 		Object result = null;
-		if(aspect.before(target, method, args)){
+		if (aspect.before(target, method, args)) {
 			try {
-//				result = ReflectUtil.invoke(target, method, args);
+				// result = ReflectUtil.invoke(target, method, args);
 				result = proxy.invokeSuper(obj, args);
-			}catch (UtilException e) {
+			} catch (UtilException e) {
 				final Throwable cause = e.getCause();
-				if(e.getCause() instanceof InvocationTargetException) {
-					aspect.afterException(target, method, args, ((InvocationTargetException)cause).getTargetException());
-				}else {
-					throw e;//其它异常属于代理的异常，直接抛出
+				if (e.getCause() instanceof InvocationTargetException) {
+					aspect.afterException(target, method, args, ((InvocationTargetException) cause).getTargetException());
+				} else {
+					throw e;// 其它异常属于代理的异常，直接抛出
 				}
 			}
 		}
-		if(aspect.after(target, method, args)){
+		if (aspect.after(target, method, args)) {
 			return result;
 		}
 		return null;