+++ hutool-aop/src/main/java/cn/hutool/aop/interceptor/JdkInterceptor.java
@@ -9,15 +9,16 @@ import cn.hutool.core.exceptions.UtilException;
 import cn.hutool.core.util.ReflectUtil;
 
 /**
- * 切面抽象
+ * JDK实现的动态代理切面
+ * 
  * @author Looly
  *
  */
-public class JdkInterceptor implements InvocationHandler{
-	
+public class JdkInterceptor implements InvocationHandler {
+
 	private Object target;
 	private Aspect aspect;
-	
+
 	/**
 	 * 构造
 	 * 
@@ -28,32 +29,32 @@ public class JdkInterceptor implements InvocationHandler{
 		this.target = target;
 		this.aspect = aspect;
 	}
-	
-	public Object getTarget(){
+
+	public Object getTarget() {
 		return this.target;
 	}
-	
+
 	@Override
 	public Object invoke(Object proxy, Method method, Object[] args) throws Throwable {
 		final Object target = this.target;
 		final Aspect aspect = this.aspect;
 		Object result = null;
-		if(aspect.before(target, method, args)){
+		if (aspect.before(target, method, args)) {
 			try {
 				result = ReflectUtil.invoke(target, method, args);
-			}catch (UtilException e) {
+			} catch (UtilException e) {
 				final Throwable cause = e.getCause();
-				if(e.getCause() instanceof InvocationTargetException) {
-					aspect.afterException(target, method, args, ((InvocationTargetException)cause).getTargetException());
-				}else {
-					throw e;//其它异常属于代理的异常，直接抛出
+				if (e.getCause() instanceof InvocationTargetException) {
+					aspect.afterException(target, method, args, ((InvocationTargetException) cause).getTargetException());
+				} else {
+					throw e;// 其它异常属于代理的异常，直接抛出
 				}
 			}
 		}
-		if(aspect.after(target, method, args)){
+		if (aspect.after(target, method, args)) {
 			return result;
 		}
 		return null;
 	}
-	
+
 }