+++ hutool-core/src/main/java/cn/hutool/core/lang/caller/ReflectionCaller.java
@@ -6,21 +6,24 @@ package cn.hutool.core.lang.caller;
  * @author Looly
  *
  */
-@SuppressWarnings({ "deprecation", "restriction" })
+@SuppressWarnings({ "deprecation"})
 public class ReflectionCaller extends SecurityManagerCaller {
 	private static final int OFFSET = 2;
 
 	@Override
+	@SuppressWarnings("restriction")
 	public Class<?> getCaller() {
 		return sun.reflect.Reflection.getCallerClass(OFFSET + 1);
 	}
 
 	@Override
+	@SuppressWarnings("restriction")
 	public Class<?> getCallerCaller() {
 		return sun.reflect.Reflection.getCallerClass(OFFSET + 2);
 	}
 
 	@Override
+	@SuppressWarnings("restriction")
 	public Class<?> getCaller(int depth) {
 		return sun.reflect.Reflection.getCallerClass(OFFSET + depth);
 	}