+++ hutool-cron/src/main/java/cn/hutool/cron/TaskExecutorManager.java
@@ -35,7 +35,8 @@ public class TaskExecutorManager {
 		}
 		// 子线程是否为deamon线程取决于父线程，因此此处无需显示调用
 		// executor.setDaemon(this.scheduler.daemon);
-		executor.start();
+//		executor.start();
+		this.scheduler.threadExecutor.execute(executor);
 		return executor;
 	}