+++ hutool-json/src/main/java/cn/hutool/json/JSONObject.java
@@ -297,12 +297,13 @@ public class JSONObject extends JSONGetter<String> implements JSON, Map<String, 
 	 */
 	@SuppressWarnings("unchecked")
 	public <T> T toBean(Type type, boolean ignoreError) {
-		if(Object.class.equals(type)) {
-			return (T) this;
-		}
 		final Class<?> clazz = TypeUtil.getClass(type);
 		if (null == clazz) {
 			throw new IllegalArgumentException(StrUtil.format("Can not know Class of Type {} !", type));
+		}else if(Object.class.equals(clazz)) {
+			return (T) this;
+		}else if(Map.class.equals(clazz)) {
+			return Convert.convert(type, this);
 		}
 		return (T) toBean(ReflectUtil.newInstance(clazz), ignoreError);
 	}