+++ hutool-json/src/test/java/cn/hutool/json/JSONUtilTest.java
@@ -82,7 +82,7 @@ public class JSONUtilTest {
 		Price price = JSONUtil.toBean(json, Price.class);
 		Assert.assertEquals("N", price.getADT().get(0).get(0).getBookingCode().get(0));
 	}
-
+	
 	@Test
 	public void putByPathTest() {
 		JSONObject json = new JSONObject();